+++ client/src/pages/about/About.jsx
@@ -1,0 +1,1 @@
+// Quienes somos Empresa y desarrolo
