+++ client/src/pages/userPages/LoginPage.jsx
@@ -3,7 +3,7 @@ import LoginForm from "../../components/UserComponents/LoginForm";
 const LoginPage = () => {
   return (
     <div>
-      <p>Estoy en el login page</p>
+      <p>Estoy en el login page iniciar seccion</p>
       <LoginForm />
     </div>
   );
